move back to GameLoader for removing unloaded games
