+++ tests/routes/FakeGameLoader.ts
@@ -38,4 +38,7 @@ export class FakeGameLoader implements IGameLoader {
   restoreGameAt(_gameId: string, _saveId: number, _cb: (game: Game | undefined) => void): void {
     throw new Error('Method not implemented.');
   }
+  unloadEndedGames(): void {
+    throw new Error('Method not implemented.');
+  }
 }
